feat: terminado controller de Player y corregido problema con Integer

+++ src/main/java/com/tommyrot/papelcesto/dto/update/UpdatePlayer.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class UpdatePlayer {
 
-  private Long playerId;
+  private Integer playerId;
   private String name;
   private String performanceId;
